+++ changelog.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - SAP: Changed how kundekontakter are created with 0 or 1 aftaler.
 
+### Added
+
+- Eflyt: Added functions for sending letters.
+
 ## [2.13.0] - 2025-08-13
 
 ### Added